+++ kotlintest-assertions/src/jsMain/kotlin/io/kotlintest/matchers/string/matchers.kt
@@ -1,0 +1,4 @@
+package io.kotlintest.matchers.string
+
+// Only supporting 0-9 for now, waiting for official support in Kotlin JS for `isDigit`
+actual fun Char.isDigit(): Boolean = this in '0'..'9'
