+++ kotlintest-tests/kotlintest-tests-core/src/jvmTest/kotlin/com/sksamuel/kotlintest/properties/PropertyAssertAllTest.kt
@@ -1,6 +1,11 @@
 package com.sksamuel.kotlintest.properties
 
-import io.kotlintest.matchers.*
+import io.kotlintest.matchers.gt
+import io.kotlintest.matchers.gte
+import io.kotlintest.matchers.lt
+import io.kotlintest.matchers.string.endWith
+import io.kotlintest.matchers.string.haveLength
+import io.kotlintest.matchers.string.startWith
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.assertAll
 import io.kotlintest.properties.forAll
